+++ src/NadekoBot/Modules/Music/Services/IMusicService.cs
@@ -3,7 +3,7 @@ using System.Diagnostics.CodeAnalysis;
 
 namespace NadekoBot.Modules.Music.Services;
 
-public interface IMusicService : IPlaceholderProvider
+public interface IMusicService
 {
     /// <summary>
     ///     Leave voice channel in the specified guild if it's connected to one
@@ -24,7 +24,7 @@ public interface IMusicService : IPlaceholderProvider
     Task EnqueueDirectoryAsync(IMusicPlayer mp, string dirPath, string queuer);
     Task<IUserMessage?> SendToOutputAsync(ulong guildId, EmbedBuilder embed);
     Task<bool> PlayAsync(ulong guildId, ulong voiceChannelId);
-    Task<IList<(string Title, string Url)>> SearchVideosAsync(string query);
+    Task<IList<(string Title, string Url, string Thumbnail)>> SearchVideosAsync(string query);
     Task<bool> SetMusicChannelAsync(ulong guildId, ulong? channelId);
     Task SetRepeatAsync(ulong guildId, PlayerRepeatType repeatType);
     Task SetVolumeAsync(ulong guildId, int value);